% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedonde.R
\name{dedondee}
\alias{dedondee}
\title{De donde esta? Shows location of IP address}
\usage{
dedondee(ip = "72.14.192.0")
}
\arguments{
\item{ip}{vector of IP address as character}
}
\value{
Leaflet map
}
\description{
Shows geographic location of given IP-address
}
\note{
El functione que dijabo el tu locasion.
}
\examples{
\dontrun{dedondee(ip="72.14.192.0")}
}
\references{
See citation("gisfin")
}
\author{
Jani Miettinen \email{jani.k.miettinen@gmail.com}
}
\keyword{utilities}
