% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{download_data}
\alias{download_data}
\title{Download data to a location}
\usage{
download_data(url, output, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{url}{A string url to data to be downloaded.}

\item{output}{A string path defining the output name.}

\item{overwrite}{Logical indicating if the output should be overwritten if
found (default: FALSE)}

\item{verbose}{logical. Should R report extra information on progress?
(default: TRUE)}
}
\value{
Logical indicating whether the download and potential extraction
operations were succesful (TRUE) or not (FALSE).
}
\description{
Archived (zips) data items will also be automatically extracted. Whehter or
not a target file is considered a zip-file is infered from the file suffix.
}
\author{
Joona Lehtomaki, Juuso Parkkinen \email{louhos@googlegroups.com}
}

