% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GeoserverHelFi.R
\docType{data}
\name{GeoserverHelFiWFSRequest}
\alias{GeoserverHelFiWFSRequest}
\title{Builds a WFS request to the \code{hel.fi} geospatial API.}
\description{
Builds a WFS request to the \code{hel.fi} geospatial API. The request is submitted with
\code{\link{GeoserverHelFiWFSClient}} to retrieve the data.
}
\section{Currently supported data sets}{

\itemize{
  \item Regional map (seutukartta)
  \item Register of public areas in Helsinki (yleisten alueiden rekisteri)
}
}

\section{Methods}{

\itemize{
  \item\code{getSeutukarttaLayers()}: A request for a list of available regional map data sets (layers).
  \item\code{getSeutukartta(layer)}: A request for regional map data \code{layer}.
  \item\code{getHelLayers()}: A request for a list of available register of public areas data sets (layers).
  \item\code{getHel(layer)}: A request for register of public areas data \code{layer}.
  \item\code{getHKRLayers()}: A request for a list of available register of public areas data sets (layers).
  \item\code{getHKR(layer)}: A request for register of public areas data \code{layer}.#' }
}
\examples{
# See the vignette.
}
\author{
Juuso Parkkinen and Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{GeoserverHelFiWFSClient}} \code{\link{WFSStreamingRequest}}
}
\keyword{datasets}

