% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ips.R
\name{get_ips}
\alias{get_ips}
\title{Get locations of IP addresses}
\usage{
get_ips(ip)
}
\arguments{
\item{ip}{vector of IP address as character}
}
\value{
Data frame with parameters ip, latitude and longitude
}
\description{
Get geographic coordinates for given IP-addresses from \url{http://www.datasciencetoolkit.org//ip2coordinates/}
}
\note{
Modified from original version by Kay Cichini and Leo Laine
}
\examples{
\dontrun{get_ips(c("137.224.252.10", "127.124.152.15"))}
}
\references{
See citation("gisfin")
}
\author{
Jani Miettinen \email{jani.k.miettinen@gmail.com}
}
\keyword{utilities}
