% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_turku_adminboundaries}
\alias{get_turku_adminboundaries}
\title{Retrieve administrative boundaries in Turku}
\usage{
get_turku_adminboundaries(map.specifier = NULL, data.dir = tempdir(),
  verbose = TRUE, ...)
}
\arguments{
\item{map.specifier}{A string. Specify the name of the Turku administrative
data set to retrieve.}

\item{data.dir}{A string. Specify a temporary folder for storing downloaded
data (default: \code{tempdir()}.}

\item{verbose}{logical. Should R report extra information on progress?
(default: TRUE)}

\item{...}{other arguments  passed on to \code{download_data}}
}
\value{
a spatial object (from SpatialPolygonsDataFrame class)
}
\description{
Retrieves administrative boundaries data from Lounaispaikka (GIS repository
for southwestern Finland).
}
\details{
Available data sets (defined by the argument \code{map.specifier}) and the
associated licenses are:

\describe{
 \item{\code{elections}}{Election districts in City of Turku (City of turku,
   \href{http://paikkatietokeskus.lounaispaikka.fi/cms/files/Avoin_data_Turku_lisenssiehdot.pdf}{Turun kaupungin lisenssi})}
 \item{\code{major-districts}}{Major district in City of Turku (City of turku,
   \href{http://paikkatietokeskus.lounaispaikka.fi/cms/files/Avoin_data_Turku_lisenssiehdot.pdf}{Turun kaupungin lisenssi})}
}
}
\examples{
# See available options
get_turku_adminboundaries()
\dontrun{
# Get the election districts as a SpatialPolygonsDataFrame
get_turku_adminboundaries('elections')

# Use cache to avoid downloading the data each time
get_turku_adminboundaries('elections', data.dir='cache')

# Force update the cache
get_turku_adminboundaries('elections', data.dir='cache', overwrite=TRUE)
}
}
\author{
Joona Lehtomaki, Juuso Parkkinen,and Leo Lahti
\email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\url{http://paikkatietokeskus.lounaispaikka.fi/fi/aineistot/},
\code{\link{download.data}}
}

