% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KarttaHelFi.R
\docType{data}
\name{KarttaHelFiWFSClient}
\alias{KarttaHelFiWFSClient}
\title{Retrieves geospatial data from \code{kartta.hel.fi}.}
\value{
In case the service at \code{hel.fi} cannot be reached, the relevant methods return \code{character(0)}.
}
\description{
Retrieves geospatial data from (\url{http://kartta.hel.fi}). A request object
to retrieve data is constructed from the class \code{\link{KarttaHelFiWFSRequest}}. Layer lists are
returned as \code{character} vectors and map data (layers) as \code{Spatial*} objects.
}
\examples{
# See the vignette.
}
\author{
Juuso Parkkinen and Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{KarttaHelFiWFSRequest}} \code{\link{WFSStreamingClient}}
}
\keyword{datasets}

