% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KarttaHelFi.R
\docType{data}
\name{KarttaHelFiWFSRequest}
\alias{KarttaHelFiWFSRequest}
\title{Builds a WFS request to the \code{kartta.hel.fi} geospatial API.}
\description{
Builds a WFS request to the \code{kartta.hel.fi} geospatial API. The request is submitted with
\code{\link{KarttaHelFiWFSClient}} to retrieve the data.
}
\section{Currently supported data sets}{

\itemize{
  \item Avoin data (avoindata)
}
}

\section{Methods}{

\itemize{
  \item\code{getAvoindataLayers()}: A request for a list of available regional map data sets (layers).
  \item\code{getAvoindata(layer)}: A request for regional map data \code{layer}.
}
}
\examples{
# See the vignette.
}
\author{
Juuso Parkkinen and Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{KarttaHelFiWFSClient}} \code{\link{WFSStreamingRequest}}
}
\keyword{datasets}

