% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GeoserverHelFi.R
\docType{data}
\name{GeoserverHelFiWFSClient}
\alias{GeoserverHelFiWFSClient}
\title{Retrieves geospatial data from \code{stat.fi}.}
\value{
In case the service at \code{stat.fi} cannot be reached, the relevant methods return \code{character(0)}.
}
\description{
Retrieves geospatial data from Statistics Finland (\url{http://www.stat.fi}). A request object
to retrieve data is constructed from the class \code{\link{GeoserverHelFiWFSRequest}}. Layer lists are
returned as \code{character} vectors and map data (layers) as \code{Spatial*} objects.
}
\examples{
# See the vignette.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("gisfin")
}
\seealso{
\code{\link{GeoserverHelFiWFSRequest}} \code{\link{WFSStreamingClient}}
}
\keyword{datasets}

